Change: Move git subprocess into common
